admin_panel/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.0
playwright==1.40.0
pydantic==2.5.0
sqlmodel==0.0.14
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
google-generativeai>=0.3.0
